+++ src/@types/react-native-pixel-perfect/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'react-native-pixel-perfect';
